+++ CHANGELOG.md
@@ -10,3 +10,4 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add class `NumPHP\Core\NumArray`
 - Add function `NumPHP\Core\NumArray::getData`
 - Add function `NumPHP\Core\NumArray::getShape`
+- Add function `NumPHP\Core\NumArray::getSize`